Remove Direct Linking to COM from DATA for `extractvars` Job (#3379)
In the current GFS version, some jobs directly write to COMOUT via
links in the working directories. This approach is risky, as downstream
jobs might receive incomplete files and fail. In order to resolve this
issue direct linking to COM from DATA should be removed. This PR
addresses this issue in the extractvars job.

Resolves #3376 
Refs #712 
Refs #3302 
Refs #3273