Add Rocoto Jobs for the Missing Products of GEFS (#3466)
<!--
  *** PLEASE READ ***

  Any PRs not following this template will be closed.

  Please delete all these comments before submitting the PR.

Please use a short (<60 char), descriptive title for the PR title above.
It should complete the sentence "If merged, this PR will _____".
Capitalize the first word and do not end with a period.

  No content should appear above the "Description" header.

If this PR is not merge-ready (e.g. it depends on other PRs not yet
merged), please mark it as draft until it is ready.

- This PR adds the rocoto jobs for gempak and bufrsnd missing products in the GEFS
system and ensure they are properly included in the XML configuration
generation. 

Refs #3469 #3470 #3467