Create separate gists when multiple files are published on GitHub (#3551)
Creates multiple gists when several files are being published at once:

This PR updates the python code that publishes gists so that it will
create multiple gists for each file that it is processing when there are
more than one error log associated with a fail. Sometime when too many
files where included in a single gist it made it difficult for the user
to find the specific file and the gist would get too large to view them
with in the comment stream.

Resolves #3542