Update wxflow to enforce failure when linking required files. (#3623)
This updates wxflow to enable the new link_req method and updates `stage_ic` yamls to use `link_req`.  It also uses `find_upward` from `wxflow` instead of `find_homegfs`.

See wxflow PRs 
https://github.com/NOAA-EMC/wxflow/pull/53 
https://github.com/NOAA-EMC/wxflow/pull/54 
https://github.com/NOAA-EMC/wxflow/pull/55
https://github.com/NOAA-EMC/wxflow/pull/57

In addition, this adds a couple lines to `.gitignore`
- `global_control.nml.IN` is a soft-link from ufs-weather-model. It
should be ignored.
- `rocotostat.out` is a file created by `build_compute.sh` to monitor
the builds. It should be ignored.

Lastly, this updates the location of the soft-link to be
consistent with the other templates linked from the ufs-weather-model.

Fixes #3547