Convert err_chk to err_exit (#3703)
This adopts a standard method of using err_exit in all J-Jobs and
ex-scripts. Simultaneously, all ush scripts had err_exit and err_chk
removed in favor of exit statements to allow ex-scripts to handle the
final error checking (with the exception of ush scripts that define
functions used directly in ex-scripts or J-Jobs).

Some cleanup was also performed to a) better handle errors, 2) conform
to EE2 standards, and 3) conform to shellcheck best practices.

Resolves https://github.com/NOAA-EMC/global-workflow/issues/3678
Resolves #3709 
Refs #2349
Resolves #294