Update marine LETKF halo size (#3759)
JEDI LETKF "halo" size for the observation distribution has to be at
least as large as the largest localization lengthscale for the algorithm
to work correctly. This is not currently the case on develop for marine
LETKF. This PR fixes the halo size to be correct.

Currently on develop this incorrect behavior still leads to succeeding
runs (with incorrect results), but once
https://github.com/JCSDA-internal/ufo/pull/3710 is merged (and
propagated back up to gdasapp and gw), this incorrect behavior will
become a runtime error in LETKF and the tests and experiments will start
to fail without the change in this PR (that's a good failure though,
alerting us to the incorrect behavior!)


Co-authored-by: Anna Shlyaeva <anna.v.shlyaeva@noaa.gov>