Updates for code hand-off, land orog sfc_data consistency check, etc. (#3892)
This
- moves `load_fv3gfs_modules.sh`, `load_ufswm_modules.sh`, and
`load_ufsda_modules.sh` from `ush/` to `dev/ush/` as these scripts are
not intended for use in ops and hence should be moved to `dev/` space.
- Fixes #3664 - adds control for wave gempak output generation and caps
it at 180h.
- Fixes #3838 - adds a checker to the forecast job. Presently shows a
warning, but a fatal error facility is added waiting to be enabled. The
checker can be run as a standalone script.
- Fixes #3825
- Fixes #3893 - checks for first half cycle if it exists and updates the
dependency tree for cycling. Tested via `C96C48mx500_S2SW_cyc_gfs` on
WCOSS2.
- Removes a section of TODO from `build_upp.sh` that depended on UFSWM.
The UFSWM dependency was satisfied.