Update the maximum number of gempak tasks to 42 (#3884)
# Description
Following up on the meeting between @ChristopherHill-NOAA
@DavidHuber-NOAA and @aerorahul and the notes from
[this](https://github.com/NOAA-EMC/global-workflow/issues/3630#issuecomment-3084414209),
this PR:
- increases the maximum tasks allowed for gempak jobs to 42 (from 25)

This will reduce the number of calls to `${NAGRIB}` made in the job.

@ChristopherHill-NOAA will investigate the performance issues with
calling too many `${NAGRIB}` calls and why `${GEMEXE}/gpend` does not
clear any residual `${NAGRIB}` processes or memory fragments.

This PR is not a solution, but a fix to mask the root cause.