Increase C96 gcdas fcst wall time to 1.5 hours (#4108)
This PR increases the wall time for the C96 gcdas (gcafs da) forecast to
`01:30:00` on Ursa.

This is done due to highly variable Ursa job run times. The GCAFS
forecast takes an unusually long time to complete on Ursa. Developers
are aware of this and are working on a long term solution. Increasing
the job wall time is a patch until the model run time stabilizes and is
in line with the run time on other machines.

Resolves #4107